+++ Bomberman/src/Model/Movable.java
@@ -3,49 +3,49 @@ package Model;
 public class Movable {
 	
 	//physical attributes
-	private int xval, yval;
+	private float xval, yval;
 	private int height, width;
 	private int speed;
 	
 	//setters
-	public void setXval(int i){
+	public void setXval(float i) {
 		xval = i;
 	}
-	public void setYval(int i){
+	public void setYval(float i) {
 		yval = i;
 	}
-	public void setHeight(int i){
+	public void setHeight(int i) {
 		height = i;
 	}
-	public void setWidth(int i){
+	public void setWidth(int i) {
 		width = i;
 	}
-	public void setSpeed(int i){
+	public void setSpeed(int i) {
 		speed = i;
 	}
 	
 	//increment
-	public void incrementXval(int i){
+	public void incrementXval(float i) {
 		xval += i;
 	}
-	public void incrementYval(int i){
+	public void incrementYval(float i) {
 		yval += i;
 	}
 
 	//getters
-	public int getXval(){
-		return xval;
+	public int getXval() {
+		return (int)xval;
 	}
-	public int getYval(){
-		return yval;
+	public int getYval( ) {
+		return (int)yval;
 	}
-	public int getHeight(){
+	public int getHeight() {
 		return height;
 	}
-	public int getWidth(){
+	public int getWidth() {
 		return width;
 	}
-	public int getSpeed(){
+	public int getSpeed() {
 		return speed;
 	}
 }
